fps/animation subsystem integration start

+++ typeft.h
@@ -15,13 +15,13 @@
 
 typedef signed char				t_int8;
 typedef signed short			t_int16;
-typedef signed long int			t_int32;
-typedef signed long long int	t_int64;
+typedef signed int				t_int32;
+typedef signed long int			t_int64;
 
 typedef unsigned char			t_uint8;
 typedef unsigned short			t_uint16;
-typedef unsigned long int		t_uint32;
-typedef unsigned long long int	t_uint64;
+typedef unsigned int			t_uint32;
+typedef unsigned long int		t_uint64;
 
 typedef unsigned char			t_uchar;
 typedef unsigned short			t_ushort;
